HappyTaps-FindTaps/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
Flask==2.1.2
google-api-core==2.7.3
google-auth==2.6.6
google-cloud==0.34.0
google-cloud-core==2.3.0
google-cloud-datastore==2.5.1
google-cloud-trace==1.6.1
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.4
grpcio==1.46.0
grpcio-status==1.46.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
opentelemetry-api==1.11.1
opentelemetry-exporter-gcp-trace==1.3.0
opentelemetry-instrumentation==0.30b1
opentelemetry-instrumentation-flask==0.30b1
opentelemetry-instrumentation-requests==0.30b1
opentelemetry-instrumentation-wsgi==0.30b1
opentelemetry-propagator-gcp==1.3.0
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
opentelemetry-util-http==0.30b1
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
rsa==4.8
six==1.16.0
slack-bolt==1.13.1
slack-sdk==3.16.1
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
